Add simple app that uses just NRF MDK
The application uses the SysTick timer to toggle the LED.

The application is built from scratch without using an example, just to
get a feel on how things work, and understand what's provided in the
mdk. The startup file, linker scripts are used directly from the MDK
without any updates.

Since, the nordic header files also include the CMSIS, the defines to
access the processor peripherals are also available.

Nordic abstracts the platform and header definitions are platform
independent, but compile time device definitions are required.

The C standard lib is not used for now.